refactor: revert unnecessary change
